requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
aiohttp
ngrok